refactor decorator controller as part of further understanding
This commit adds code level comments, renames a few variables as
part of deeper understanding of decorator controller. This has
been done to reduce the steep learning curve required by contributors
to understand and use this controller/library.

This is part of a series of commits that will lead to implementing
unit tests as well re-packaging dynamic package to enable using
the same as a library in other projects.

Signed-off-by: AmitKumarDas <amit.das@mayadata.io>