refactor hook logic (#10)
This commit tries to decouple hook invocation logic from the
schema that defines these hooks. With this commit hook invocation
logic relies only on the fields needed to invoke the hook. For
example, a web hook invocation relies only on the URL and timeout
to invoke the hook and not on the entire hook schema.

This decoupling has been attempted by making use of following
programming principles:
1/ use of functional options extensively.
2/ moving away from free flowing functions to corresponding
structures.

This is a good to have modification, since it makes newer hook
implementation de-coupled from their schema definition. Some of
the future hooks that are being considered are:

1/ config based hooks where the logic to arrive at the desired
state can be in Kubernetes based configmaps.

2/ inline hooks where logic can be invoked directly by the binary
that uses metac as a dependent library.

Signed-off-by: AmitKumarDas <amit.das@mayadata.io>