[usecase, gctl]: showcase install/un-install of CRD
This commit showcases how GenericController can be used to install
and uninstall a CRD when a particular namespace gets created and
deleted correspondingly.

This commit can be understood in two ways:
1/ Files in test/integration/generic/ shows how golang based
integration code can be written to verify this use-case.

2/ Files in examples/gctl/install-uninstall-crd/ shows how metac
controller & jsonnet files are used to verify this use-case.

NOTES:
1/ There are some changes to vendoring as well. Latest k8s.io
dependencies are set.

2/ Integration test case has been set to run only if test is not
run in short mode.

Signed-off-by: AmitKumarDas <amit.das@mayadata.io>