add skipReconcile to GenericController hook response
SkipReconcile is a flag that skips reconciliation of
GenericController's attachments. This tunable is only supported
for GenericController.

SkipReconcile looks similar to GenericController's spec.ReadOnly.
However, both of them serve different purposes. SkipReconcile is
dynamic and is set by the sync hook implementation. However ReadOnly
is a static option that is set in GenericController's spec.

In other words, one expects SkipReconcile to vary from true to false
or vice versa depending on runtime conditions. On the other hand,
ReadOnly is expected to be set to true or false for the entire
lifecycle of the controller.

Signed-off-by: AmitKumarDas <amit.das@mayadata.io>