[hook] add support to invoke inline functions as hooks
This commit adds support to specify inline functions as meta
controller hooks. In other words, metacontroller code as well
as hook implementation will be built as a single binary. This
avoids the need to have http service based invocations.

In addition to inline hook support, this commit has following changes:
- code refactor w.r.t hook invocation
- a new integration test case that showcases use of inline hook

NOTE: This support is also enabled for GenericController. Once
this proves to be useful, it may be extended to other meta controllers.

Signed-off-by: AmitKumarDas <amit.das@mayadata.io>