add generated client code to let metac be used as a lib (#37)
This commit adds the generated client code as versioned source
code. This helps in letting other projects make use of metac
as a library.

In addition, an integration test is skipped when tests are being
run in short mode. This was done since this particular integration
test was very flaky.

Signed-off-by: AmitKumarDas <amit.das@mayadata.io>