feat(patch): add patch support to GenericController
This commit adds optional patch support to merge operations
in GenericController. This enables controller to remove any
field values of a resource even if this resource was not
created by this controller. Without this a controller can
only add or update values to above resource.

In addition, this change has several unit tests that verifies
metac's 3-way merge operations. Unit tests have been added to
several packages that participate in this merge operation.

Signed-off-by: AmitKumarDas <amit.das@mayadata.io>