feat(gctl): add explicit update & delete logic to generic controller
This commit adds logic to update or delete attachments that were
not created by the generic controller. In other words, a controller
developer can control updating or deleting any attachment via a
generic controller even if this attachment was not created by this
controller.

This feature was already available in GenericController via UpdateAny
and DeleteAny flags. However, these flags were too generic and enabled
update or delete of any kind of attachment. With this new change a
developer can selectively update or delete specific attachment
instances.

This change has also some amount of variable and logging modifications
to make the code as well as logs readable.

Signed-off-by: AmitKumarDas <amit.das@mayadata.io>