feat(integration-test): add integration test framework
This commit removes the old way to implement an integration test.
This introduces a new approach to test any feature in a better
& readable format. Any integration test can be implemented as a
series of steps. Each of this step can either be a create or update
or delete or assert of any kubernetes resource.

In addition generic, composite as well as decorator tests are clubbed
into a single package. Integration tests are still divided into two
categories (i.e. crd or config mode) based on the way metac binary
can be started.

There will be more enhancements to integration tests going forward
which in turn will align metac to be production grade inspite of
new features or changes to existing ones.

Signed-off-by: AmitKumarDas <amit.das@mayadata.io>